+++ lang.txt
@@ -1,0 +1,6 @@
+E̶n̶g̶l̶i̶s̶h̶
+F̶r̶a̶n̶ç̶a̶i̶s̶
+P̶o̶r̶t̶u̶g̶u̶ê̶s̶
+E̶s̶p̶a̶ñ̶o̶l̶
+Deutsch
+Italiano
